xml load moved to siteservice
